Fix error in IE8- for empty elements
When an element has no content, `elem.innerText` will be falsey,
so the right hand side of the `||` operator is used.
Since `.textContent` is `undefined` in IE8 and before, the result
will not be a string, but `undefined`.
Later operations which except the variable to be a string will fail,
because `undefined` does not support any string operations.